requirements.txt
openai>=1.0.0
click>=8.1.7
python-dotenv>=1.0.1
